functions/requirements.txt
firebase-functions
firebase-admin
requests
python-dateutil